+++ src/curves/mod.rs
@@ -13,13 +13,14 @@
 //!
 //! ## Terminology
 //!
-//! We distinguish between two kinds of curves. The most basic curves ([`Sin`],
-//! [`SawTri`], [`Pulse`], etc.) are all examples of (plain) **curves**, meaning
-//! types implementing [`Map<Input = f32, Output = f32>`](Map).
+//! We distinguish between two kinds of curves. The most basic curves like
+//! [`Sin`], [`SawTri`], [`Pulse`], etc.) are all examples of
+//! **(plain) curves**, meaning types implementing
+//! [`Map<Input = f64, Output = f64>`](Map).
 //!
 //! On the other hand, audio curves such as those generated from
 //! [`BufCurve`](crate::prelude::BufCurve) are called **sample curves**, meaning
-//! types implementing [`Map<Input = f32>`](Map)` where Output: Sample`.
+//! types implementing [`Map<Input = f64>`](Map)` where Output: Sample`.
 
 pub mod buffer;
 
@@ -38,15 +39,15 @@ impl Pos {
 }
 
 impl Map for Pos {
-    type Input = f32;
-    type Output = f32;
+    type Input = f64;
+    type Output = f64;
 
-    fn eval(&self, x: f32) -> f32 {
+    fn eval(&self, x: f64) -> f64 {
         crate::pos(x)
     }
 }
 
-impl<F: Map<Output = f32>> Comp<F, Pos> {
+impl<F: Map<Output = f64>> Comp<F, Pos> {
     /// Composes a function with [`Pos`].
     pub const fn pos(f: F) -> Self {
         Self::new(f, Pos)
@@ -66,15 +67,15 @@ impl Sgn {
 }
 
 impl Map for Sgn {
-    type Input = f32;
-    type Output = f32;
+    type Input = f64;
+    type Output = f64;
 
-    fn eval(&self, x: f32) -> f32 {
+    fn eval(&self, x: f64) -> f64 {
         crate::sgn(x)
     }
 }
 
-impl<F: Map<Output = f32>> Comp<F, Sgn> {
+impl<F: Map<Output = f64>> Comp<F, Sgn> {
     /// Composes a function with [`Sgn`].
     pub const fn sgn(f: F) -> Self {
         Self::new(f, Sgn)
@@ -94,15 +95,15 @@ impl Neg {
 }
 
 impl Map for Neg {
-    type Input = f32;
-    type Output = f32;
+    type Input = f64;
+    type Output = f64;
 
-    fn eval(&self, x: f32) -> f32 {
+    fn eval(&self, x: f64) -> f64 {
         -x
     }
 }
 
-impl<F: Map<Output = f32>> Comp<F, Neg> {
+impl<F: Map<Output = f64>> Comp<F, Neg> {
     /// Composes a function with [`Neg`].
     pub const fn neg(f: F) -> Self {
         Self::new(f, Neg)
@@ -113,33 +114,33 @@ impl<F: Map<Output = f32>> Comp<F, Neg> {
 #[derive(Clone, Copy, Debug)]
 pub struct Linear {
     /// Slope of the map.
-    pub slope: f32,
+    pub slope: f64,
 
     /// y-intercept of the map.
-    pub intercept: f32,
+    pub intercept: f64,
 }
 
 impl Linear {
     /// Initializes a new linear map.
     #[must_use]
-    pub const fn new(slope: f32, intercept: f32) -> Self {
+    pub const fn new(slope: f64, intercept: f64) -> Self {
         Self { slope, intercept }
     }
 
     /// Initializes the linear map that rescales an interval
     /// `[init_lo, init_hi]` to another `[end_lo, end_hi]`.
     #[must_use]
-    pub fn rescale(init_lo: f32, init_hi: f32, end_lo: f32, end_hi: f32) -> Self {
+    pub fn rescale(init_lo: f64, init_hi: f64, end_lo: f64, end_hi: f64) -> Self {
         let slope = (end_hi - end_lo) / (init_hi - init_lo);
         Self::new(slope, end_lo - slope * init_lo)
     }
 }
 
 impl Map for Linear {
-    type Input = f32;
-    type Output = f32;
+    type Input = f64;
+    type Output = f64;
 
-    fn eval(&self, x: f32) -> f32 {
+    fn eval(&self, x: f64) -> f64 {
         x * self.slope + self.intercept
     }
 }
@@ -165,10 +166,10 @@ impl Saw {
 }
 
 impl Map for Saw {
-    type Input = f32;
-    type Output = f32;
+    type Input = f64;
+    type Output = f64;
 
-    fn eval(&self, x: f32) -> f32 {
+    fn eval(&self, x: f64) -> f64 {
         crate::sgn(x)
     }
 }
@@ -194,10 +195,10 @@ impl InvSaw {
 }
 
 impl Map for InvSaw {
-    type Input = f32;
-    type Output = f32;
+    type Input = f64;
+    type Output = f64;
 
-    fn eval(&self, x: f32) -> f32 {
+    fn eval(&self, x: f64) -> f64 {
         -crate::sgn(x)
     }
 }
@@ -222,10 +223,10 @@ impl PosSaw {
 }
 
 impl Map for PosSaw {
-    type Input = f32;
-    type Output = f32;
+    type Input = f64;
+    type Output = f64;
 
-    fn eval(&self, x: f32) -> f32 {
+    fn eval(&self, x: f64) -> f64 {
         x
     }
 }
@@ -250,10 +251,10 @@ impl PosInvSaw {
 }
 
 impl Map for PosInvSaw {
-    type Input = f32;
-    type Output = f32;
+    type Input = f64;
+    type Output = f64;
 
-    fn eval(&self, x: f32) -> f32 {
+    fn eval(&self, x: f64) -> f64 {
         1.0 - x
     }
 }
@@ -264,13 +265,13 @@ impl Map for PosInvSaw {
 #[derive(Clone, Copy, Debug, Default)]
 pub struct Sin {
     /// Phase of the sine curve, from `0.0` to `1.0`.
-    pub phase: f32,
+    pub phase: f64,
 }
 
 impl Sin {
     /// A sine wave with a given phase.
     #[must_use]
-    pub const fn new(phase: f32) -> Self {
+    pub const fn new(phase: f64) -> Self {
         Self { phase }
     }
 
@@ -290,15 +291,15 @@ impl Sin {
 
 /// Evaluates `sin((x + phase) * τ)`.
 #[must_use]
-pub fn sin(x: f32, phase: f32) -> f32 {
-    ((x + phase) * std::f32::consts::TAU).sin()
+pub fn sin(x: f64, phase: f64) -> f64 {
+    ((x + phase) * std::f64::consts::TAU).sin()
 }
 
 impl Map for Sin {
-    type Input = f32;
-    type Output = f32;
+    type Input = f64;
+    type Output = f64;
 
-    fn eval(&self, x: f32) -> f32 {
+    fn eval(&self, x: f64) -> f64 {
         sin(x, self.phase)
     }
 }
@@ -309,13 +310,13 @@ impl Map for Sin {
 #[derive(Clone, Copy, Debug)]
 pub struct Pulse {
     /// Shape of the pulse curve, from `0.0` to `1.0`.
-    pub shape: f32,
+    pub shape: f64,
 }
 
 impl Pulse {
     /// A pulse wave.
     #[must_use]
-    pub const fn new(shape: f32) -> Self {
+    pub const fn new(shape: f64) -> Self {
         Self { shape }
     }
 
@@ -334,7 +335,7 @@ impl Default for Pulse {
 
 /// Returns `1` if `x < shape`, returns `-1` otherwise.
 #[must_use]
-pub fn pulse(x: f32, shape: f32) -> f32 {
+pub fn pulse(x: f64, shape: f64) -> f64 {
     if x < shape {
         1.0
     } else {
@@ -343,10 +344,10 @@ pub fn pulse(x: f32, shape: f32) -> f32 {
 }
 
 impl Map for Pulse {
-    type Input = f32;
-    type Output = f32;
+    type Input = f64;
+    type Output = f64;
 
-    fn eval(&self, x: f32) -> f32 {
+    fn eval(&self, x: f64) -> f64 {
         pulse(x, self.shape)
     }
 }
@@ -357,13 +358,13 @@ impl Map for Pulse {
 #[derive(Clone, Copy, Debug, Default)]
 pub struct SawTri {
     /// Position of the peak of the wave.
-    pub shape: f32,
+    pub shape: f64,
 }
 
 impl SawTri {
     /// A saw-triangle curve.
     #[must_use]
-    pub const fn new(shape: f32) -> Self {
+    pub const fn new(shape: f64) -> Self {
         Self { shape }
     }
 
@@ -395,9 +396,9 @@ impl SawTri {
 /// Interpolates linearly between `-1.0` and `1.0` for `x ≤ shape`, and between
 /// `1.0` and `-1.0` for `x ≥ shape`.
 #[must_use]
-pub fn saw_tri(x: f32, shape: f32) -> f32 {
+pub fn saw_tri(x: f64, shape: f64) -> f64 {
     /// We must do some size checks to avoid division by 0.
-    const EPS: f32 = 1e-7;
+    const EPS: f64 = 1e-7;
 
     if x < shape {
         if shape.abs() < EPS {
@@ -413,10 +414,10 @@ pub fn saw_tri(x: f32, shape: f32) -> f32 {
 }
 
 impl Map for SawTri {
-    type Input = f32;
-    type Output = f32;
+    type Input = f64;
+    type Output = f64;
 
-    fn eval(&self, x: f32) -> f32 {
+    fn eval(&self, x: f64) -> f64 {
         saw_tri(x, self.shape)
     }
 }
